renamed helper folder to helperVariables
