+++ crudOperations/quoteOperations/quoteModel.js
@@ -1,6 +1,6 @@
 const db = require("../../databaseOperations/db-config");
 const axios = require("axios");
-const { headers } = require("../helpers/headers");
+const { headers } = require("../helperVariables/headers");
 
 module.exports = {
   insert,
